Move `extract_byte_ranges_read[_seek]` from `array::codec` to `storage::byte_range`
